backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic==2.3.0
requests==2.31.0
psutil==5.9.5
huggingface-hub==0.16.4
aiohttp==3.8.5
python-multipart==0.0.6
tqdm==4.66.1
gitpython==3.1.40
pillow==10.0.0
# Optional GPU monitoring dependencies
py3nvml==0.2.7
gputil==1.4.0
